Added AMR to convergence instructions, as it does not harm if we have

See merge request hpcsoftware/Peano!281

Former-commit-id: fd4cbadd8c6f57c1089cc44a95674f1746373d6c

+++ python/exahype2/postprocessing/plot-scaling.py
@@ -16,6 +16,42 @@ Colors = [ "#aa0000", "#00aa00", "#0000aa", "#797900", "#790079", "#007979", "#a
 Symbols = [ "o", "s", "<", ">" , "v", "^" ]
 
 
+
+def visualise( data_points ):
+    global single_node_time 
+    global max_nodes        
+    global max_time         
+    
+    (x_data, y_data) = exahype2.postprocessing.extract_times_per_step( data_points, args.max_cores_per_rank )    
+     
+    if len(x_data)>0:
+      max_nodes = max(max_nodes,x_data[-1])
+    
+    if args.plot_efficiency:
+      normalised_fasted_time = y_data[0] * x_data[0]
+      for i in range(0,len(x_data)):
+        y_data[i] = normalised_fasted_time / y_data[i] / x_data[i]
+      if args.max_cores_per_rank>0:
+        y_data = [y/float(args.max_cores_per_rank) for y in y_data]
+      y_data = [ min(y,1.1) for y in y_data]
+      
+    symbol = "-" + Symbols[ args.file.split(",").index(file) % len(Symbols) ]
+    my_markevery = 0.1 + 0.8 * (args.file.split(",").index(file)/len(args.file.split(",")))**3
+    my_color = Colors[ args.file.split(",").index(file) % len(Colors) ]
+    
+    max_time = max(y_data[0], max_time)
+    
+    if args.labels=="":
+      plt.plot( x_data, y_data, symbol, color=my_color, markevery=my_markevery )
+    else:
+      try:
+        my_label = args.labels.split(",")[ args.file.split(",").index(file) ]
+      except:
+        raise Exception( "Unable to extract " + str(args.file.split(",").index(file)) + "th entry from " + args.labels + ": " + str(args.labels.split(",")) )
+      plt.plot( x_data, y_data, symbol, label=my_label, color=my_color, markevery=my_markevery )
+    
+
+
 if __name__ == "__main__":
   parser = argparse.ArgumentParser(description="""
 ExaHyPE 2 scaling plotter:
@@ -23,7 +59,7 @@ ExaHyPE 2 scaling plotter:
 A generic script to create speedup plots.
 
 """)
-  parser.add_argument("file",   help="filename of the file to parse (should be a tar.gz file). If you have multiple files, seperate them with a comma")
+  parser.add_argument("file",   help="filename of the file to parse (should be a tar.gz file or a directory). If you have multiple files, seperate them with a comma")
   parser.add_argument("-v",                  "--verbose", help="increase output verbosity", action="store_true" )
   parser.add_argument("--max-cores-per-rank", dest="max_cores_per_rank", type=int, help="max number of cores per rank (pick 0 if you have only one core count per rank, pick -1 if you want to plot single-node data)", default=0)
   parser.add_argument("--log-x",              dest="log_x",           help="plot with logarithmic axes", action="store_true" )
@@ -41,57 +77,50 @@ A generic script to create speedup plots.
   plt.clf()
   
   for file in args.file.split(","):
-    open_mode = ""  
-    if file.endswith( "tar.gz"):
-      open_mode = "r:gz"
-    elif file.endswith( "tar" ):
-      open_mode = "r:"
-    else:
-      print( "Warning: passed file should end with tar.gz or tar" )
+    if os.path.isdir(file):
+      print( "Parse file " + file )
+      data_points = []
+      for data_file in os.listdir(file):
+        print( "========================================================================")
+        print( file + "/" + data_file + " from " + file )
+        print( "========================================================================")
+        new_data = exahype2.postprocessing.PerformanceData(file + "/" + data_file, args.verbose)
+        if new_data.valid:
+          data_points.append( new_data ) 
+      visualise( data_points )
+    elif file.endswith( "tar") or file.endswith("tar.gz"):
+      print( "Parse archive " + file )
+      open_mode = ""  
+      if file.endswith( "tar.gz"):
+        open_mode = "r:gz"
+      elif file.endswith( "tar" ):
+        open_mode = "r:"
 
-    tar = tarfile.open( file, "r:gz" )
-    data_files = tar.getnames()
-
-    data_points = []
-    for data_file in data_files:
-      print( "========================================================================")
-      print( data_file + " from " + file )
-      print( "========================================================================")
-      tar.extract( data_file )
-      new_data = exahype2.postprocessing.PerformanceData(data_file, args.verbose)
-      if new_data.valid:
-        data_points.append( new_data ) 
-      os.remove( data_file )
-
-    (x_data, y_data) = exahype2.postprocessing.extract_times_per_step( data_points, args.max_cores_per_rank )    
-     
-    if len(x_data)>0:
-      max_nodes = max(max_nodes,x_data[-1])
-    
-    if args.plot_efficiency:
-      normalised_fasted_time = y_data[0] * x_data[0]
-      for i in range(0,len(x_data)):
-        y_data[i] = normalised_fasted_time / y_data[i] / x_data[i]
-      if args.max_cores_per_rank>0:
-        y_data = [y/float(args.max_cores_per_rank) for y in y_data]
-      y_data = [ min(y,1.1) for y in y_data]
-      
-    symbol = "-" + Symbols[ args.file.split(",").index(file) % len(Symbols) ]
-    my_markevery = 0.1 + 0.8 * (args.file.split(",").index(file)/len(args.file.split(",")))**3
-    my_color = Colors[ args.file.split(",").index(file) % len(Colors) ]
-
-    max_time = max(y_data[0], max_time)
-    
-    if args.labels=="":
-      plt.plot( x_data, y_data, symbol, color=my_color, markevery=my_markevery )
-    else:
+      tar = None
       try:
-        my_label = args.labels.split(",")[ args.file.split(",").index(file) ]
-      except:
-        raise Exception( "Unable to extract " + str(args.file.split(",").index(file)) + "th entry from " + args.labels + ": " + str(args.labels.split(",")) )
-      plt.plot( x_data, y_data, symbol, label=my_label, color=my_color, markevery=my_markevery )
-      
-  
+        tar = tarfile.open( file, "r:gz" )
+        data_files = tar.getnames()
+    
+        data_points = []
+        for data_file in data_files:
+          print( "========================================================================")
+          print( data_file + " from " + file )
+          print( "========================================================================")
+          tar.extract( data_file )
+          new_data = exahype2.postprocessing.PerformanceData(data_file, args.verbose)
+          if new_data.valid:
+            data_points.append( new_data ) 
+          os.remove( data_file )
+          
+        visualise( data_points )
+      except Exception as e:
+        print( "Error: " + str(e) )
+        tar = None
+      if tar!=None:   
+        tar.close()
+    else:
+      print( file + " is neither a directory or an archive" )
+
   if args.plot_efficiency:
     plt.ylabel( "Efficiency" )
   else:
@@ -117,5 +146,3 @@ A generic script to create speedup plots.
       output_file_name = filename
   plt.savefig( output_file_name + ".pdf" )
   plt.savefig( output_file_name + ".png" )
-  
-  tar.close()
